mini: Fix some pointer/integer cast warnings on 32-bit.

+++ mono/mini/unwind.c
@@ -709,7 +709,7 @@ mono_unwind_frame (guint8 *unwind_info, guint32 unwind_info_len,
 		mono_dump_mem (ip - 0x10, 0x40);
 		return FALSE;
 	}
-	cfa_val = (guint8*)regs [mono_dwarf_reg_to_hw_reg (cfa_reg)] + cfa_offset;
+	cfa_val = (guint8*)(gsize)regs [mono_dwarf_reg_to_hw_reg (cfa_reg)] + cfa_offset;
 	for (hwreg = 0; hwreg < NUM_HW_REGS; ++hwreg) {
 		if (reg_saved [hwreg] && locations [hwreg].loc_type == LOC_OFFSET) {
 			int dwarfreg = mono_hw_reg_to_dwarf_reg (hwreg);
